+++ src/lib/types/employees.types.ts
@@ -1,0 +1,22 @@
+import { create } from 'zustand';
+import { Timestamp } from 'firebase/firestore';
+
+export interface Employee {
+  id?: string;
+  id_company: string;
+  company_name: string;
+  id_role: string;
+  role_name: string;
+  id_departament: string;
+  departament_name: string;
+  leader: boolean;
+  name: string;
+  cpf: string;
+  sex: string;
+  birth_date: string;
+  scholarity: string;
+  admission_date: string;
+  login: string;
+  password: string;
+  created_at: Timestamp;
+}
