+++ src/shared/providers.tsx
@@ -1,6 +1,16 @@
-import { SidebarProvider } from '@/components/ui/sidebar';
+'use client';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactNode } from 'react';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+
+// Create a client
+const queryClient = new QueryClient();
 
 export default function Providers({ children }: { children: ReactNode }) {
-  return <SidebarProvider>{children}</SidebarProvider>;
+  return (
+    <QueryClientProvider client={queryClient}>
+      {children}
+      <ReactQueryDevtools initialIsOpen={false} />
+    </QueryClientProvider>
+  );
 }
